move open method into base
